+++ lifelines/fitters/weibull_fitter.py
@@ -29,7 +29,7 @@ class WeibullFitter(KnownModelParametericUnivariateFitter):
     Important
     ----------
     The parameterization of this model changed in lifelines 0.19.0. Previously, the cumulative hazard looked like
-    :math:`(\lambda t)^\rho`. The parameterization is now the recipricol of :math:`\lambda`.
+    :math:`(\lambda t)^\rho`. The parameterization is now the reciprocal of :math:`\lambda`.
 
     Examples
     --------